refact: apple 로그인 메소드 분리
